refactor: 사용자 글쓰기 제한 알림 기록 저장 로직 수정 * 저장 전에 모든 글쓰기 제한 알림 기록 삭제 * SaveNotifcationHisotry의 메소드 NotificationUseCase로 이동
