+++ boot/external-api/src/main/java/com/sooum/api/card/service/CardLikeService.java
@@ -1,6 +1,6 @@
 package com.sooum.api.card.service;
 
-import com.sooum.api.notification.service.SaveNotificationService;
+import com.sooum.api.notification.service.NotificationUseCase;
 import com.sooum.api.notification.service.SendFCMService;
 import com.sooum.data.card.entity.CommentCard;
 import com.sooum.data.card.entity.CommentLike;
@@ -11,6 +11,7 @@ import com.sooum.data.card.service.CommentLikeService;
 import com.sooum.data.card.service.FeedCardService;
 import com.sooum.data.card.service.FeedLikeService;
 import com.sooum.data.member.entity.Member;
+import com.sooum.data.member.entity.devicetype.DeviceType;
 import com.sooum.data.member.service.MemberService;
 import jakarta.persistence.EntityExistsException;
 import lombok.RequiredArgsConstructor;
@@ -27,7 +28,7 @@ public class CardLikeService {
     private final FeedLikeService feedLikeService;
     private final CommentLikeService commentLikeService;
     private final MemberService memberService;
-    private final SaveNotificationService saveNotificationService;
+    private final NotificationUseCase notificationUseCase;
     private final SendFCMService sendFCMService;
 
     @Transactional
@@ -51,7 +52,7 @@ public class CardLikeService {
                         .build()
         );
 
-        saveNotificationService.saveFeedLikeHistory(requesterPk, targetCard);
+        notificationUseCase.saveFeedLikeHistory(requesterPk, targetCard);
     }
 
     @Transactional
@@ -84,7 +85,8 @@ public class CardLikeService {
                 .targetCard(targetCard)
                 .build());
 
-        saveNotificationService.saveCommentLikeHistory(requesterPk, targetCard);
+        notificationUseCase.saveCommentLikeHistory(requesterPk, targetCard);
+        sendFCMService.sendCommentLikeMsg(DeviceType.ANDROID, targetFeedCardPk, targetCard.getWriter().getNickname(), "c78DeTomQF6yjOy86mcJ11:APA91bGtHV5-17g6u8d7vVH_LrYtqoJiVgmlMpHG3ahV9wMteIw2I_77jJviBWxPXWjXEC8W3BAcNQNqZaEZn6eErsbKTIMuMMP8aH88UCkK1QEl3mtAtcc");
     }
 
     @Transactional
